+++ rosen/modules/render_service/composer/composer_base/connection/rs_render_to_composer_connection_stub.cpp
@@ -91,10 +91,10 @@ int32_t RSRenderToComposerConnectionStub::GetCleanLayerBufferSurfaceId(OHOS::Mes
     return COMPOSITOR_ERROR_OK;
 }
 
-std::set<uint32_t> RSRenderToComposerConnectionStub::ParseClearRedrawCacheBufferIds(OHOS::MessageParcel& parcel)
+std::set<uint64_t> RSRenderToComposerConnectionStub::ParseClearRedrawCacheBufferIds(OHOS::MessageParcel& parcel)
 {
-    std::vector<uint32_t> bufferIdsVector;
-    std::set<uint32_t> bufferIds;
+    std::vector<uint64_t> bufferIdsVector;
+    std::set<uint64_t> bufferIds;
     if (!parcel.ReadUInt32Vector(&bufferIdsVector)) {
         RS_LOGE("RSRenderToComposerConnectionStub::ParseClearRedrawCacheBufferIds read buffer ids failed.");
         return bufferIds;
